catalog: carve TableService out of the facade
Move table CRUD (register/register_from_data/overwrite/get/list/
update/delete), reference resolution (resolve_table, resolve_table_out,
resolve_write_destination, resolve_table_file_path), DTO conversion
(table_to_out, bulk_enrich_tables), worker materialisation, file-existence
checks, schema parsing, and table favourites into
catalog/services/tables.py.

Apply Category A bare-except cleanup at _read_table_metadata
(narrowed from ``except Exception`` to ``(RuntimeError, OSError)``)
and Category C cleanup in _compute_laziness_blockers (now uses
resolve_or_log).

register_table_from_parquet body stays on the facade — a test
monkeypatches CatalogService.register_table_from_data and expects
the alias to route through it. _materialize_table_with_worker uses
a lazy module lookup of trigger_catalog_materialize so test
monkeypatches on catalog.service.trigger_catalog_materialize keep
working.

Re-export CatalogMaterializationResult on CatalogService for the
test that references CatalogService.CatalogMaterializationResult.
